New module label
Move label related stuff to new module "label". Also changed PrimedLabel
class to the more generic name Label.
