+++ tncontract/__init__.py
@@ -17,5 +17,6 @@ twodimensional
 
 from tncontract.version import __version__
 from tncontract.tensor import *
+from tncontract.label import *
 import tncontract.onedim
 import tncontract.twodim
