+++ tncontract/tensor.py
@@ -1,12 +1,12 @@
 import warnings
 import numpy as np
 import scipy as sp
-import uuid
 
 __all__ = ['Tensor', 'contract', 'distance', 'matrix_to_tensor',
         'tensor_to_matrix', 'random_tensor', 'tensor_product', 'tensor_svd',
-        'truncated_svd', 'unique_label', 'zeros_tensor', 'prime_label',
-        'unprime_label', 'prime_level']
+        'truncated_svd', 'zeros_tensor']
+
+from tncontract import label as lbl
 
 class Tensor():
     """
@@ -114,7 +114,7 @@ class Tensor():
         for i, label in enumerate(self.labels):
             for noprime_label in labels:
                 if noprime_label(label) == noprime_label:
-                    self.labels[i] = prime_label(self.labels[i])
+                    self.labels[i] = lbl.prime_label(self.labels[i])
 
     def unprime_label(self, labels):
         """
@@ -141,7 +141,7 @@ class Tensor():
         for i, label in enumerate(self.labels):
             for noprime_label in labels:
                 if noprime_label(label) == noprime_label:
-                    self.labels[i] = unprime_label(self.labels[i])
+                    self.labels[i] = lbl.unprime_label(self.labels[i])
 
     def contract_internal(self, label1, label2, index1=0, index2=0):
         """By default will contract the first index with label1 with the 
@@ -278,75 +278,6 @@ class Tensor():
     def shape(self):
         return self.data.shape
 
-
-class PrimedLabel(str):
-    """Wrapper class for priming string-like labels"""
-
-    def __new__(cls, value, parent=None):
-        return str.__new__(cls, value)
-
-    def __init__(self, label, parent=None):
-        self._parent = parent
-
-    @property
-    def parent(self):
-        """Return parent label"""
-        return self._parent
-
-    @property
-    def origin(self):
-        """Return origin label"""
-        origin = self
-        while hasattr(origin, "parent"):
-            origin = origin.parent
-        return origin
-
-    @property
-    def parents(self):
-        """Return number of parents for label"""
-        tmp = self
-        level = 0
-        while hasattr(tmp, "parent"):
-            if tmp.parent is not None:
-                tmp = tmp.parent
-                level += 1
-            else:
-                break
-        return level
-
-def prime_label(label, prime="'"):
-    """Put a prime on a label object"""
-    return PrimedLabel(label+prime, parent=label)
-
-def unprime_label(label, prime="'"):
-    """Remove one prime from label object"""
-    try:
-        parent = label.parent
-    except AttributeError:
-        raise ValueError("label is not primed")
-    if parent+prime == label:
-        return parent
-    else:
-        raise ValueError("label is not primed with \"" + prime + "\"")
-
-def noprime_label(label):
-    """Remove all primes from a label object"""
-    try:
-        return label.origin
-    except AttributeError:
-        return label
-
-def prime_level(label):
-    """Return number of primes on label object"""
-    try:
-        return label.parents
-    except AttributeError:
-        return 0
-
-
-def unique_label():
-    """Generate a long, random string that is very likely to be unique."""
-    return str(uuid.uuid4())
 
 #Tensor constructors
 def random_tensor(*args, labels=[], base_label="i"):
